update P2P synchronisation
the P2P features of Zero Tours only work in one direction at a time, so it won't do to only receive as many packets as are sent; in fact, it seems that this was never the issue with Pokémon. rather, the issue with Pokémon was that it was receiving a low-level error on remote disconnect, which isn't even a detectable state when the connection is over a phone line. rather, we simply stop sending the game data, allowing it to detect the remote disconnect through a timeout or other method which, to reiterate, was necessary in '01/'02 when these games were connecting over a phone line.